Slightly reorder AuthenticationController
